rename the fitting and plotting to be :tools:
