+++ src/qicklab/analysis/starkspec.py
@@ -4,7 +4,7 @@ import numpy as np
 from ..utils.ana_utils  import rotate_and_threshold
 from ..utils.data_utils import process_h5_data
 from ..utils.file_utils import load_from_h5_with_shotdata
-from .plotting import plot_stark_simple
+from .plot_tools import plot_stark_simple
 from .shot_tools import process_shots
 from .stark_tools import gain2freq_detuning
